Refactor divide file for maintenance
